requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
assertpy==1.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
curlify==2.2.1
distconfig3==1.0.1
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pydantic==2.7.4
pydantic_core==2.18.4
PyHamcrest==2.1.0
pytest==8.2.2
PyYAML==6.0.1
requests==2.32.3
retrying==1.3.4
six==1.16.0
structlog==24.2.0
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.1
vyper-config==1.2.1
watchdog==4.0.1